樣式修正

+++ Web/Services/CourseService.cs
@@ -701,7 +701,7 @@ namespace Web.Services
                 select new CourseReview
                 {
                     ReviewerName = member.FirstName + " " + member.LastName,
-                    CommentRating = comment.Rating,
+                    CommentRating = (int)comment.Rating,
                     ReviewDate = comment.Cdate.ToString("yyyy/MM/dd"),
                     ReviewContent = comment.CommentText
                 }).ToListAsync();
